Extracted GoogleApi auth project; added CI workflow #6
