+++ .github/workflows/google-api-auth-ci.yml
@@ -1,0 +1,22 @@
+name: GoogleApi.Authorization build
+
+on: 
+  pull_request:
+    branches:
+      - master
+    paths:
+      - 'GoogleApi/src/GoogleApi.Authorization/**'
+
+jobs:
+  build:
+    runs-on: ubuntu-latest
+    steps:
+      - uses: actions/checkout@v2
+      - name: setup .NET
+        uses: actions/setup-dotnet
+        with:
+          dotnet-version: 3.1.201
+      - name: dotnet restore
+        run: dotnet restore
+      - name: build
+        run: dotnet build --configuration Release --no-restore
