+++ GoogleApi/src/GmailApi.Client/GmailApi.Client.csproj
@@ -14,4 +14,8 @@
     <PackageReference Include="Google.Apis.Gmail.v1" Version="1.42.0.1777" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\GoogleApi.Authorization\GoogleApi.Authorization.csproj" />
+  </ItemGroup>
+
 </Project>
