+++ GoogleApi/src/GoogleApi.sln
@@ -7,7 +7,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GoogleApi.Console", "Google
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GmailApi.Client", "GmailApi.Client\GmailApi.Client.csproj", "{6CBF3DF0-9C9B-45B8-ACF3-3AB911A529C1}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GmailApi.Client.Tests", "GmailApi.Client.Tests\GmailApi.Client.Tests.csproj", "{ADEFCAD1-F4FF-4A6D-99AD-74D18E734AE4}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "GmailApi.Client.Tests", "GmailApi.Client.Tests\GmailApi.Client.Tests.csproj", "{ADEFCAD1-F4FF-4A6D-99AD-74D18E734AE4}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GoogleApi.Authorization", "GoogleApi.Authorization\GoogleApi.Authorization.csproj", "{D15A7ABC-0C4B-44E4-B01A-491B98A39245}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -27,6 +29,10 @@ Global
 		{ADEFCAD1-F4FF-4A6D-99AD-74D18E734AE4}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{ADEFCAD1-F4FF-4A6D-99AD-74D18E734AE4}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{ADEFCAD1-F4FF-4A6D-99AD-74D18E734AE4}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D15A7ABC-0C4B-44E4-B01A-491B98A39245}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D15A7ABC-0C4B-44E4-B01A-491B98A39245}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D15A7ABC-0C4B-44E4-B01A-491B98A39245}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D15A7ABC-0C4B-44E4-B01A-491B98A39245}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
